refactor(api): move tooling/env.ts to top-level env.ts
Single file in a single-purpose directory—no reason for the nesting.
Both consumers (better-auth.config.ts, drizzle.config.ts) are top-level
config files themselves.
