+++ apps/api/drizzle.config.ts
@@ -1,7 +1,7 @@
 import { fileURLToPath } from 'node:url';
 import { config } from 'dotenv';
 import { defineConfig } from 'drizzle-kit';
-import { LOCAL_DATABASE_URL } from './tooling/env';
+import { LOCAL_DATABASE_URL } from './env';
 
 config({ path: fileURLToPath(new URL('.dev.vars', import.meta.url)) });
